requirements.txt
fastapi
uvicorn
stripe
google-generativeai
sqlalchemy
python-jose[cryptography]
python-multipart